+++ packages/components/bolt-video/main.js
@@ -1,2 +1,0 @@
-import './src/video.standalone.js';
-import './src/video-meta.js';